+++ examples/data-sources/influxdb3_cloud_database/data-source.tf
@@ -1,0 +1,3 @@
+data "influxdb3_cloud_database" "signals" {
+  name = "signals"
+}