+++ examples/data-sources/influxdb3_cloud_database/outputs.tf
@@ -1,0 +1,3 @@
+output "signals_database" {
+  value = data.influxdb3_cloud_database.signals
+}