+++ examples/data-sources/influxdb3_cloud_databases/data-source.tf
@@ -1,0 +1,1 @@
+data "influxdb3_cloud_databases" "all" {}