+++ examples/data-sources/influxdb3_cloud_token/data-source.tf
@@ -1,0 +1,3 @@
+data "influxdb3_cloud_token" "signals_token" {
+  id = "7f7fa77d-b77e-77ba-7777-77cd077d0f7c"
+}