+++ examples/data-sources/influxdb3_cloud_tokens/data-source.tf
@@ -1,0 +1,1 @@
+data "influxdb3_cloud_tokens" "all" {}