+++ examples/data-sources/influxdb3_cloud_tokens/outputs.tf
@@ -1,0 +1,4 @@
+output "all_tokens" {
+  value     = data.influxdb3_cloud_tokens.all.tokens
+  sensitive = true
+}