+++ examples/resources/influxdb3_cloud_database/outputs.tf
@@ -1,0 +1,3 @@
+output "signals_database" {
+  value = influxdb3_cloud_database.signals
+}