+++ examples/resources/influxdb3_cloud_token/import.sh
@@ -1,0 +1,2 @@
+# Tokens are imported by ID.
+terraform import influxdb3_cloud_token.signals 7f7fa77d-b77e-77ba-7777-77cd077d0f7c